+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The sliding behavior in `o-slider` component requires that HammerJS is loaded on
 ## 3.2.1
 ### Features
 * **o-form**: new `getFormComponentPermissions` and  `getActionsPermissions` methods
+* **o-table**: new attribute `select-all-checkbox-visible` ([a553447](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/a553447))
 
 ### BREAKING CHANGES
 * **PermissionsService**: ([2d186c9](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/2d186c9))
@@ -44,6 +45,8 @@ The sliding behavior in `o-slider` component requires that HammerJS is loaded on
 * **o-grid**: pagination local in grid  ([916e632](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/916e632)), closes [#229](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/229)
 * **o-table**: expand the table in the container
   ([8d50a68](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/8d50a68))
+* **o-grid** and **o-table**: emit changes event in translations of the page
+  ([f627826](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/f627826))
 * **ontimize-export.service**: fixing download bug ([dbeaf42](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/dbeaf42)) [#237](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/237)
 
 ## 3.2.0 (2018-11-27)